move di to module
