+++ app/build.gradle
@@ -40,6 +40,8 @@ android {
 
 dependencies {
 
+    implementation project(path:':dil')
+
     implementation 'com.squareup.retrofit2:adapter-rxjava3:2.9.0'
 
     implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
