+++ app/src/main/java/ru/gb/appgitusers/di/Module.kt
@@ -9,11 +9,11 @@ import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
 import ru.gb.appgitusers.data.UserRepo
 import ru.gb.appgitusers.data.retrofit.GitUsersApi
-import ru.gb.appgitusers.data.room.GitUserDao
 import ru.gb.appgitusers.data.room.GitUserDataBase
 import ru.gb.appgitusers.domain.IGitUserRepository
+import ru.gb.dil.DependenciesHolder
 
-class Module(contex:Context, diHolder: DependenciesHolder) {
+class Module(contex: Context, diHolder: DependenciesHolder) {
     private val baseUrl = "https://api.github.com/"
     private val nameDb = "GitUsers.db"
     private val datasource by lazy {
@@ -42,8 +42,9 @@ class Module(contex:Context, diHolder: DependenciesHolder) {
     val userRepo: IGitUserRepository by lazy {
         //LocalGitUserRepository()
         //ApiGitUserRepository()
-        UserRepo(api,datasource)
+        UserRepo(api, datasource)
     }
+
     init {
         diHolder.add(UserRepo::class, userRepo)
     }
